Move import/export functions to api
