+++ ui/src/KaptureApp/api/importGrids.js
@@ -7,63 +7,10 @@ import {
 import {
   COMPONENT_TYPE_ATTRIBUTE,
   createComponent,
-  exportComponent,
 } from 'KaptureApp/config/componentTypes';
 import { GRID_ROW_HEADERS } from 'KaptureApp/config/grid';
 
-const exportGrids = (grids) => {
-  const exportedGrids = grids.map((grid, i) => {
-    return exportGrid(grid, i + 1);
-  });
-  return exportedGrids;
-};
-
-const exportGrid = (grid, id) => {
-  const exportedGrid = {
-    id,
-    rows: grid.dimensions.rows,
-    columns: grid.dimensions.columns,
-    name: grid.name || null,
-    containerType: grid.containerType || null,
-    barcode: grid.barcode || null,
-    data: [],
-  };
-  if (grid.data && grid.data.length) {
-    const positions = grid.data.flat();
-    positions.forEach((position) => {
-      if (position.container) {
-        exportedGrid.data.push(
-          exportContainer(position.container, position.row, position.column)
-        );
-      }
-    });
-  }
-  return exportedGrid;
-};
-
-const exportContainer = (container, row, column) => {
-  const exportedContainer = {
-    name: container.name || null,
-    containerType: container.containerType || null,
-    row: row || null,
-    col: column || null,
-    barcode: container.barcode || null,
-    components: [],
-    attributes: [],
-  };
-  if (container.components && container.components.length) {
-    container.components.forEach((component) => {
-      if (component.type === COMPONENT_TYPE_ATTRIBUTE) {
-        exportedContainer.attributes.push(exportComponent(component));
-      } else {
-        exportedContainer.components.push(exportComponent(component));
-      }
-    });
-  }
-  return exportedContainer;
-};
-
-const importGrids = (grids, kaptureComponents) => {
+export const importGrids = (grids, kaptureComponents) => {
   const importData = {
     grids: [],
     barcodes: [],
@@ -77,7 +24,7 @@ const importGrids = (grids, kaptureComponents) => {
   return importData;
 };
 
-const importGrid = (importData, kaptureComponents) => {
+function importGrid(importData, kaptureComponents) {
   const gridData = createGridData(
     {
       rows: importData.rows,
@@ -100,9 +47,9 @@ const importGrid = (importData, kaptureComponents) => {
   });
   addContainersToGrid(grid, containerPositions, GRID_ROW_HEADERS);
   return grid;
-};
+}
 
-const importContainer = (containerData, kaptureComponents) => {
+function importContainer(containerData, kaptureComponents) {
   const editorComponents = containerData.components.map((component) => {
     const kaptureComponent = kaptureComponents[component.type].find(
       (kaptureComponent) => kaptureComponent.id === component.id
@@ -132,6 +79,4 @@ const importContainer = (containerData, kaptureComponents) => {
     barcode: containerData.barcode,
     components: containerComponents,
   });
-};
-
-export { exportGrids, importGrids };
+}
